@@ -6,7 +6,7 @@ version: 2.5.1
 homepage: https://github.com/tneotia/html-editor-enhanced
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <=4.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
